third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/averagePool2d.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] averagePool2d float32 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
[FAIL] averagePool2d float32 4D tensor options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 50.03894805908203 should be close enough to expected 21.206613540649414 by ULP distance: expected a number less than or equal to 11n but got 10387645n
[FAIL] averagePool2d float32 4D tensor options.roundingType=ceil and no padding
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 50.63130187988281 should be close enough to expected 40.29140853881836 by ULP distance: expected a number less than or equal to 11n but got 2710541n
[FAIL] averagePool2d float32 4D tensor options.layout=nhwc and options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 50.03894805908203 should be close enough to expected 21.206613540649414 by ULP distance: expected a number less than or equal to 11n but got 10387645n
[FAIL] averagePool2d float32 4D tensor options.outputSizes ignores options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 50.03894805908203 should be close enough to expected 21.206613540649414 by ULP distance: expected a number less than or equal to 11n but got 10387645n
[FAIL] averagePool2d float16 4D constant tensor all positive default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 47.28125 by ULP distance: expected a number less than or equal to 27 but got 20969
[FAIL] averagePool2d float16 4D tensor all positive default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 47.28125 by ULP distance: expected a number less than or equal to 27 but got 20969
[FAIL] averagePool2d float16 4D tensor all negative default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected -49.25 by ULP distance: expected a number less than or equal to 27 but got 53800
[FAIL] averagePool2d float16 4D tensor options.windowDimensions
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 43.46875 by ULP distance: expected a number less than or equal to 11 but got 20847
[FAIL] global averagePool2d float16 4D tensor all positive options.windowDimensions
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 47.28125 by ULP distance: expected a number less than or equal to 27 but got 20969
[FAIL] averagePool2d float16 4D tensor options.padding
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 52.4375 by ULP distance: expected a number less than or equal to 27 but got 21134
[FAIL] averagePool2d float16 4D tensor options.strides
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 43.46875 by ULP distance: expected a number less than or equal to 11 but got 20847
[FAIL] averagePool2d float16 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
[FAIL] averagePool2d float16 4D tensor options.layout=nchw
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 47.28125 by ULP distance: expected a number less than or equal to 27 but got 20969
[FAIL] averagePool2d float16 4D tensor options.layout=nhwc
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 47.28125 by ULP distance: expected a number less than or equal to 27 but got 20969
[FAIL] global averagePool2d float16 4D tensor options.layout=nhwc and options.windowDimensions
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 47.28125 by ULP distance: expected a number less than or equal to 27 but got 20969
[FAIL] averagePool2d float16 4D tensor options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 54.1875 by ULP distance: expected a number less than or equal to 11 but got 21190
[FAIL] averagePool2d float16 4D tensor options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 54.1875 by ULP distance: expected a number less than or equal to 11 but got 21190
[FAIL] averagePool2d float16 4D tensor options.roundingType=ceil and no padding
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 51.1875 by ULP distance: expected a number less than or equal to 11 but got 21094
[FAIL] averagePool2d float16 4D tensor options.layout=nhwc and options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 54.1875 by ULP distance: expected a number less than or equal to 11 but got 21190
[FAIL] averagePool2d float16 4D tensor options.layout=nhwc and options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 54.1875 by ULP distance: expected a number less than or equal to 11 but got 21190
[FAIL] averagePool2d float16 4D tensor options.outputSizes ignores options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 54.1875 by ULP distance: expected a number less than or equal to 11 but got 21190
[FAIL] averagePool2d float16 4D tensor options.outputSizes ignores options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 54.1875 by ULP distance: expected a number less than or equal to 11 but got 21190
[FAIL] averagePool2d float16 4D tensor options.dilations with options.strides
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float16 actual 0 should be close enough to expected 42.9375 by ULP distance: expected a number less than or equal to 11 but got 20830
Harness: the test ran to completion.
